refactor: deduplicate QtAudioPlayer/QtVideoPlayer's worker-thread
New in qtplayer.py:
- WorkerThreadPlayer: the shared Player subclass __init__ wiring and  _engine_* implementations. Subclasses set _worker_cls and override _extra_setup/_connect_extra/_on_load_success/_extra_exit_cleanup hooks for anything beyond it (QtVideoPlayer's widget/FrameRelay). _requests_cls defaults to EngineRequests since audio and video have no reason to diverge there.
- EngineWorker/EngineRequests: the same treatment one level down, for audio_worker.py/video_worker.py's worker-thread classes. AudioEngineRequests/VideoEngineRequests were literally identical, so both now just use EngineRequests directly rather than two empty subclasses. EngineWorker's _extra_worker_setup/_extra_worker_exit_cleanup hooks are named distinctly from WorkerThreadPlayer's GUI-thread-side hooks of a similar shape, to avoid two same-named hooks with different thread affinity living in unrelated classes.
- Completion moved here too (was identically copy-pasted in both worker modules).

qtaudio.py/qtvideo.py and audio_worker.py/video_worker.py now hold only what's genuinely audio- or video-specific: qtaudio.py is a 16-line subclass; video_worker.py keeps the QVideoSink pairing and the frame-relay backpressure timer. VideoEngineWorker/AudioEngineWorker are siblings of EngineWorker, not one subclassing the other -- video isn't a kind of audio, they're two independent uses of the same underlying QMediaPlayer/worker-thread machinery, and video inheriting from something named Audio would be a misleading hierarchy that also risks silently absorbing audio-only behaviour if AudioEngineWorker ever needs any.

Also fixed real type-annotation bugs in base.py's abstract _engine_* methods, surfaced while reviewing the seam between qtplayer.py and the base class: _engine_load_media was declared -> None but every implementation returns bool and load_media() actually consumes that value; _engine_exit was declared -> float for no reason anything ever needed; _engine_stop had no return annotation at all despite one implementation returning an unused value. Propagated the same corrections to CLIYoutubePlayer's stubs and YouTubePlayer, and
tightened wait_for_signal's own typing with ParamSpec so the decorator preserves the wrapped function's actual parameters instead of collapsing them to *args/**kwargs: Any. _worker_cls/_requests_cls are now type[EngineWorker]/type[EngineRequests] instead of bare type.

CI caught a real bug this introduced: WorkerThreadPlayer.__init__ constructed QMediaDevices() before calling _extra_setup(), reversing QtVideoPlayer's original order (its QVideoWidget used to be built first). That inversion caused a real, reproducible crash (Fatal Python error: Aborted, inside long_operation.py's processEvents() pump) when
loading a local video -- confirmed by bisecting against this same commit with and without the reorder, not just inferred. Fixed by running _extra_setup() before QMediaDevices() again, restoring the original order for QtVideoPlayer (a no-op for QtAudioPlayer, which has
no widget).
