+++ CLAUDE.md
@@ -75,7 +75,7 @@ Once a `QWebEngineView` exists in-process (the YouTube player in `tilia/media/pl
 - `tilia/ui/webengine_tracking.py` monkey-patches `QWebEngineView.__init__` to set a process-global, one-way flag (`any_web_engine_view_created()`) the first time any such view is constructed anywhere — catching both current call sites and future ones without per-call-site opt-in.
 - `tilia/ui/long_operation.py::_pump()` consults that flag to choose a safe fallback (`QCoreApplication.sendPostedEvents(None, QEvent.Type.LayoutRequest)` + `repaint()`) instead of `processEvents()` when showing progress feedback.
 - `tilia/media/loader.py::_change_player_type` has a related but distinct hazard: switching away from the YouTube player must flush `DeferredDelete` events before constructing the new player, to avoid a macOS CoreAudio/AVFoundation deadlock. Don't conflate the two — different mechanism, different trigger condition.
-- `tilia/media/player/qtplayer.py::wait_for_signal`'s nested `QEventLoop().exec()` was the hazard's root cause for media playback (the old `QtPlayer` class that used it directly has since been removed). **Fixed for `QtAudioPlayer`, `QtVideoPlayer`, and CLI's players** (`tilia/ui/cli/player.py::CLIVideoPlayer` and the `"audio"` mapping in `tilia/ui/cli/ui.py::get_player_class` both now reuse `QtAudioPlayer`): each runs `QMediaPlayer`/`QAudioOutput` on a worker `QThread` (`audio_worker.py`, `video_worker.py`), blocking the calling thread on a plain `threading.Event` instead of a nested event loop — a raw OS wait can't reenter the GUI thread's dispatcher. Position and (for video) frames are pushed from the worker via `PositionRelay`/`FrameRelay` rather than the GUI thread polling-and-blocking for them, which used to cause real lag under CPU contention. `QtVideoPlayer` additionally can't pair its worker-thread player with the GUI-thread `QVideoWidget` via `setVideoOutput()` directly — that looks fine (no crash, correct state transitions) but silently drops all frame delivery; only manual visual testing catches it. **Lesson: no-crash and correct-state-transitions aren't evidence of correct rendering, and a blocking round-trip belongs on one-shot calls only, never a hot polling loop.**
+- `tilia/media/player/qtplayer.py::wait_for_signal`'s nested `QEventLoop().exec()` was the hazard's root cause for media playback (the old `QtPlayer` class that used it directly has since been removed). **Fixed for `QtAudioPlayer`, `QtVideoPlayer`, and CLI's players** (`tilia/ui/cli/player.py::CLIVideoPlayer` and the `"audio"` mapping in `tilia/ui/cli/ui.py::get_player_class` both now reuse `QtAudioPlayer`): all three share `qtplayer.py`'s `WorkerThreadPlayer`/`EngineWorker`/`EngineRequests` base classes, which run `QMediaPlayer`/`QAudioOutput` on a worker `QThread` (`audio_worker.py`/`video_worker.py` hold only the video-specific `QVideoSink`/frame-relay additions — see `EngineWorker`'s `_extra_worker_*` hooks), blocking the calling thread on a plain `threading.Event` instead of a nested event loop — a raw OS wait can't reenter the GUI thread's dispatcher. Position and (for video) frames are pushed from the worker via `PositionRelay`/`FrameRelay` rather than the GUI thread polling-and-blocking for them, which used to cause real lag under CPU contention. `QtVideoPlayer` additionally can't pair its worker-thread player with the GUI-thread `QVideoWidget` via `setVideoOutput()` directly — that looks fine (no crash, correct state transitions) but silently drops all frame delivery; only manual visual testing catches it. **Lesson: no-crash and correct-state-transitions aren't evidence of correct rendering, and a blocking round-trip belongs on one-shot calls only, never a hot polling loop.**
 - Normal modal `QDialog`/`QMessageBox`/`QFileDialog`/`QMenu` `.exec()` calls elsewhere in `tilia/ui/` are **not** part of this hazard family — they're expected blocking UI, not a reentrant pump, and don't need this treatment.
 
 ## Code style
